requirements.txt
FastApi
mysql-connector-python
#-e .